# Mesa de Ayuda/MesaDeAyudaRest/src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Modulo de Mesa de Ayuda API",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8082",
        "description": "Servidor local"
      }
    ],
    "paths": {
      "/loginPost": {
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Devuelve el token de autorizacion",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "description": "Username del usuario"
                    },
                    "password": {
                      "type": "string",
                      "description": "Password del usuario"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Informacion del usuario",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "ID dependiendo del tipo de usuario"
                      },
                      "auth_token": {
                        "type": "string",
                        "description": "Token de autorizacion para el usuario"
                      },
                      "username": {
                        "type": "string",
                        "description": "Username del usuario"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/registerPost": {
        "post": {
          "tags": [
            "Users"
          ],
          "summary": "Devuelve el usuario creado",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username":{
                      "type": "string",
                      "description": "Nombre para la cuenta del usuario"
                    },
                    "password":{
                      "type": "string",
                      "description": "Contraseña elegida por el usuario para su cuenta"
                    },
                    "rolId":{
                      "type": "integer",
                      "description": "Rol elegido para la cuenta del usuario (Comprador, Vendedor)"
                    },
                    "nombre":{
                      "type": "string",
                      "description": "Nombre del usuario"
                    },
                    "apellido":{
                      "type": "string",
                      "description": "Apellido del usuario"
                    },
                    "dni":{
                      "type": "integer",
                      "description": "DNI del usuario"
                    },
                    "telefono":{
                      "type": "integer",
                      "description": "Telefono del usuario"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Informacion del usuario",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "El ID autogenerado por MySQL"
            },
            "username": {
              "type": "string",
              "description": "El nombre de usuario que indentifica al User"
            },
            "nombre": {
              "type": "string",
              "description": "El nombre del usuario"
            },
            "apellido": {
              "type": "string",
              "description": "El apellido del usuario"
            },
            "password": {
              "type": "string",
              "description": "La contraseña elegida por el usuario"
            },
            "dni": {
              "type": "integer",
              "description": "El DNI del usuario"
            },
            "telefono": {
              "type": "integer",
              "description": "El telefono del usuario"
            },
            "billetera": {
              "type": "integer",
              "description": "La cuenta del usuario"
            },
            "rol":{
              "type": "object",
              "$ref": "#/components/schemas/Rol"
            }
          }
        },
...
        "Rol": {
          "type": "object",
          "properties": {
            "id":{
              "type": "integer",
              "description": "El ID autogenerado por MySQL"
            },
            "tipo":{
              "type": "string",
              "description": "El tipo de rol asignado a cada usuario"
            }
          }
        }
    }
  }
}